First commit
First commit of my new project Yapper, which is going to be like a discord / slack application. Where users can send messages, pictures and gifs to each-other.

In this first commit I have built out LoginBox.jsx and CreateUserBox.jsx, which is basically the functionality to either create a new account or login to a previous one.

Other than that set up prisma, proxy server and added som routes for handling the requests to the back end