Landing page and conversations updates
Mostly new UI changes in LandingPage.jsx, for example the updating header. Also added some transitions for the modules so it feels a bit more engaging.

Separated the login and sign-up containers into separate pages instead of switching based on state. So now it is in LoginBox.jsx
and SignUp.jsx