Documentation and Supabase updates
Done with the documentation in the back end, however I needed to adress the file handeling situation. Previously only the file path was the unique reference, so another user could overwrite their image even if they were diffrent, just because the path was the same, so now that is adressed. Users can now also upload the same image or file several times, before it just overwrite the old one.

Next step is to also update the avatars, because right now they have the same problem